platform/masker-library/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
numpy>=1.26,<3.0
sounddevice>=0.4,<1.0
webrtcvad-wheels>=2.0.14,<3.0
faster-whisper>=1.0,<2.0
python-dotenv>=1.0,<2.0
